Insert sort the logs
Instead of appending/sorting. Since most of the time the log will be
received in order, insert sort just becomes an append.

The race condition will come when we fetch back log of event while we
already have logged a very few of event already. Since the log will be
inserted in order, it will be a small insert sort, almost O(1) cost.